refactor: 按职责拆分download_manager.rs
