+++ src-tauri/src/commands.rs
@@ -425,15 +425,15 @@ pub fn resume_download_task(app: AppHandle, chapter_uuid: String) -> CommandResu
 #[allow(clippy::needless_pass_by_value)]
 #[tauri::command(async)]
 #[specta::specta]
-pub fn cancel_download_task(app: AppHandle, chapter_uuid: String) -> CommandResult<()> {
+pub fn delete_download_task(app: AppHandle, chapter_uuid: String) -> CommandResult<()> {
     let download_manager = app.get_download_manager();
 
     download_manager
-        .cancel_download_task(&chapter_uuid)
+        .delete_download_task(&chapter_uuid)
         .map_err(|err| {
-            CommandError::from(&format!("取消章节ID为`{chapter_uuid}`的下载任务失败"), err)
+            CommandError::from(&format!("删除章节ID为`{chapter_uuid}`的下载任务失败"), err)
         })?;
-    tracing::debug!("取消章节ID为`{chapter_uuid}`的下载任务成功");
+    tracing::debug!("删除章节ID为`{chapter_uuid}`的下载任务成功");
     Ok(())
 }
