+++ src-tauri/src/downloader.rs
@@ -1,0 +1,4 @@
+pub mod download_img_task;
+pub mod download_manager;
+pub mod download_task;
+pub mod download_task_state;
