+++ src-tauri/src/downloader/download_task_state.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+use specta::Type;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Type)]
+pub enum DownloadTaskState {
+    Pending,
+    Downloading,
+    Paused,
+    Completed,
+    Failed,
+}
