+++ src-tauri/src/events.rs
@@ -5,34 +5,30 @@ use specta::Type;
 use tauri_specta::Event;
 
 use crate::{
-    download_manager::DownloadTaskState,
+    downloader::download_task_state::DownloadTaskState,
     types::{ChapterInfo, Comic, LogLevel},
 };
 
 #[derive(Debug, Clone, Serialize, Deserialize, Type, Event)]
-pub struct DownloadSpeedEvent {
-    pub speed: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Type, Event)]
-#[serde(rename_all = "camelCase")]
-pub struct DownloadSleepingEvent {
-    pub chapter_uuid: String,
-    pub remaining_sec: u64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Type, Event)]
-#[serde(rename_all = "camelCase")]
-pub struct DownloadControlRiskEvent {
-    pub chapter_uuid: String,
-    pub retry_after: u32,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, Type, Event)]
 #[serde(tag = "event", content = "data")]
-pub enum DownloadTaskEvent {
+pub enum DownloadEvent {
     #[serde(rename_all = "camelCase")]
-    Create {
+    Speed { speed: String },
+
+    #[serde(rename_all = "camelCase")]
+    RiskControl {
+        chapter_uuid: String,
+        retry_after: u32,
+    },
+
+    #[serde(rename_all = "camelCase")]
+    Sleeping {
+        chapter_uuid: String,
+        remaining_sec: u64,
+    },
+
+    #[serde(rename_all = "camelCase")]
+    TaskCreate {
         state: DownloadTaskState,
         comic: Box<Comic>,
         chapter_info: Box<ChapterInfo>,
@@ -41,7 +37,10 @@ pub enum DownloadTaskEvent {
     },
 
     #[serde(rename_all = "camelCase")]
-    Update {
+    TaskDelete { chapter_uuid: String },
+
+    #[serde(rename_all = "camelCase")]
+    TaskUpdate {
         chapter_uuid: String,
         state: DownloadTaskState,
         downloaded_img_count: u32,
