+++ src/types.ts
@@ -1,8 +1,8 @@
-import { DownloadTaskEvent } from './bindings.ts'
+import { DownloadEvent } from './bindings.ts'
 
 export type CurrentTabName = 'search' | 'favorite' | 'downloaded' | 'chapter'
 
-export type ProgressData = Extract<DownloadTaskEvent, { event: 'Create' }>['data'] & {
+export type ProgressData = Extract<DownloadEvent, { event: 'TaskCreate' }>['data'] & {
   percentage: number
   indicator: string
   retryAfter: number
